🔧 Stop using the `cgi` stdlib module
the module was deprecatied in py 3.11 and removed in py 3.13
there are examples of using the email module to resolve this:
PEP 594
https://github.com/python-babel/babel/issues/873
https://stackoverflow.com/questions/69068527/python-3-cgi-parse-header

This method doesn't seem to work for cherrypy.
The email.message module is trying to create a higher level
object that is intelligent about the headers it has defined. The
cgi.parse_header function is a very low level unintelligent function
that simply parses header content based on structure and does not
inspect the contents. Because of the intelligence of the email.message
object cherrypy can't use it in the very generic way that the
cgi.parse_header function was being used.

Fix #2014